requirements.txt
flask
elevenlabs
better-profanity
python-dotenv
waitress